ai-service/requirements.txt
matplotlib
pandas
scikit-learn
seaborn
opencv-python

Pillow

tensorflow-datasets

fastapi
uvicorn[standard]
python-multipart